"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ CSPP2-Assignments/m2/Assignment-3/Solution.java
@@ -1,7 +1,7 @@
 import java.util.Scanner;
 
 public class Solution {
-	public static int power(int base,int exponent) {
+	public static long power(int base,int exponent) {
 		if(exponent==0) {
 			return 1;
 		}
@@ -16,7 +16,7 @@ public class Solution {
 		Scanner s=new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long result=power(base,exponent);
         System.out.println(result);
 	}
 }
